Change style to return t_token_list * in all process_char files
