+++ Makefile
@@ -22,7 +22,6 @@ LIBREADLINE_DIR	= readline
 # Files
 SRC				= $(SRC_DIR)/main.c \
 					$(SRC_DIR)/lexer/constructor.c \
-					$(SRC_DIR)/lexer/delimit_input.c \
 					$(SRC_DIR)/lexer/get_next_token.c \
 					$(SRC_DIR)/lexer/getter.c \
 					$(SRC_DIR)/lexer/lexer.c \
