+++ src/lexer/lexer_internal.h
@@ -6,7 +6,7 @@
 /*   By: yliu <yliu@student.42.jp>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/30 22:25:14 by yliu              #+#    #+#             */
-/*   Updated: 2024/09/03 20:13:22 by yliu             ###   ########.fr       */
+/*   Updated: 2024/09/04 10:35:04 by yliu             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,13 +28,12 @@ typedef struct s_lexer
 }					t_lexer;
 
 void				construct_lexer(const char *input, t_lexer *lexer);
-void				delimit_input(t_lexer *lexer);
 t_token_list		*get_next_token(t_lexer *lexer);
 bool				is_start_of_input(t_lexer *lexer);
-void				process_blank(t_lexer *lexer);
-void				process_comment(t_lexer *lexer);
-void				process_eof(t_lexer *lexer);
-void				process_operator(t_lexer *lexer);
-void				process_quote(t_lexer *lexer);
+t_token_list		*process_blank(t_lexer *lexer);
+t_token_list		*process_comment(t_lexer *lexer);
+t_token_list		*process_eof(t_lexer *lexer);
+t_token_list		*process_operator(t_lexer *lexer);
+t_token_list		*process_quote(t_lexer *lexer);
 
 #endif
