+++ src/lexer/process_comment.c
@@ -6,14 +6,15 @@
 /*   By: yliu <yliu@student.42.jp>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/31 22:05:30 by yliu              #+#    #+#             */
-/*   Updated: 2024/08/31 22:55:38 by yliu             ###   ########.fr       */
+/*   Updated: 2024/09/04 10:21:20 by yliu             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "lexer_internal.h"
 
-void	process_comment(t_lexer *lexer)
+t_token_list	*process_comment(t_lexer *lexer)
 {
 	lexer->type = TOKEN_EOF;
 	lexer->value = NULL;
+	return (construct_token(lexer->type, lexer->value));
 }
